Comprobar o pedir que se compruebe si están correctas de manera teorica.

+++ src/main/java/bryan/travelAPP/repository/AirlineRepository.java
@@ -3,5 +3,5 @@ package bryan.travelAPP.repository;
 import bryan.travelAPP.domain.Airline;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AirlineRepository  extends CrudRepository<Airline, Long> {
+public interface AirlineRepository  extends CrudRepository<Airline, Integer> {
 }
